backend/requirements.txt
Django==4.2.1
django-cors-headers==4.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
Pillow==9.5.0
django-dotenv==1.4.2